requirements.txt
asgiref==3.8.1
black==24.10.0
build==1.2.2.post1
CacheControl==0.14.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cleo==2.1.0
click==8.1.7
crashtest==0.4.1
distlib==0.3.9
Django==5.1.4
django-rich==1.13.0
djangorestframework==3.15.2
dulwich==0.21.7
fastjsonschema==2.21.1
filelock==3.16.1
idna==3.10
iniconfig==2.0.0
installer==0.7.0
jaraco.classes==3.4.0
keyring==24.3.1
markdown-it-py==3.0.0
mdurl==0.1.2
more-itertools==10.5.0
msgpack==1.1.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pexpect==4.9.0
pkginfo==1.12.0
platformdirs==4.3.6
pluggy==1.5.0
poetry-core==1.9.1
poetry-plugin-export==1.8.0
psycopg==3.2.3
ptyprocess==0.7.0
pycparser==2.22
Pygments==2.18.0
pyproject_hooks==1.2.0
pytest==8.3.4
RapidFuzz==3.10.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
shellingham==1.5.4
sqlparse==0.5.2
tomlkit==0.13.2
trove-classifiers==2024.10.21.16
typing_extensions==4.12.2
urllib3==2.2.3
virtualenv==20.28.0
xattr==1.1.0
